Refactor navbar components and introduce motion animations
Updated the structure of the navbar components to include a MobileNav
component, introduced Framer Motion animations for the navbar, and
reorganized the folder structure for better organization and separation
of concerns. The changes enhance the user experience by providing
smoother animations and better component organization. No issues
referenced.
