+++ app/lib/animation/navbarVariants.ts
@@ -1,0 +1,54 @@
+export const navScreenVariant = {
+  initial: {
+    scaleY: 0,
+  },
+  animate: {
+    scaleY: 1,
+    transition: {
+      duration: 0.5,
+      ease: [0.12, 0, 0.39, 0],
+    },
+  },
+  exit: {
+    scaleY: 0,
+    transition: {
+      duration: 0.5,
+      delay: 0.5,
+      ease: [0.22, 1, 0.36, 1],
+    },
+  },
+};
+
+export const linkContainerAnimationVariants = {
+  initial: {
+    transition: {
+      staggerChildren: 0.09,
+      staggerDirection: -1,
+    },
+  },
+  animate: {
+    transition: {
+      delayChildren: 0.3,
+      staggerChildren: 0.09,
+      staggerDirection: 1,
+    },
+  },
+};
+
+export const linkAppearAnimationVariant = {
+  initial: {
+    y: "30vh",
+    transition: {
+      duration: 0.5,
+      ease: [0.37, 0, 0.63, 1],
+    },
+  },
+  animate: {
+    y: 0,
+    transition: {
+      duration: 0.7,
+      ease: [0, 0.55, 0.45, 1],
+    },
+  },
+  exit: {},
+};
