+++ app/lib/components/navbar/MobileNav.tsx
@@ -1,51 +1,19 @@
 "use client";
 import React, { useState } from "react";
 import { AnimatePresence, motion } from "framer-motion";
-import { Route, routeCollection } from "../data/route";
+import { Route, routeCollection } from "../../data/route";
 import { IonIcon } from "@ionic/react";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
+import {
+  linkAppearAnimationVariant,
+  linkContainerAnimationVariants,
+  navScreenVariant,
+} from "../../animation/navbarVariants";
 
 const MobileNav = () => {
   const [isNavOpen, setIsNavOpen] = useState(true);
   const pathname = usePathname();
-
-  const navScreenVariant = {
-    initial: {
-      scaleY: 0,
-    },
-    animate: {
-      scaleY: 1,
-      transition: {
-        duration: 0.5,
-        ease: [0.12, 0, 0.39, 0],
-      },
-    },
-    exit: {
-      scaleY: 0,
-      transition: {
-        duration: 0.5,
-        delay: 0.5,
-        ease: [0.22, 1, 0.36, 1],
-      },
-    },
-  };
-
-  const linkContainerAnimationVariants = {
-    initial: {
-      transition: {
-        staggerChildren: 0.09,
-        staggerDirection: -1,
-      },
-    },
-    animate: {
-      transition: {
-        delayChildren: 0.3,
-        staggerChildren: 0.09,
-        staggerDirection: 1,
-      },
-    },
-  };
 
   return (
     <>
@@ -120,23 +88,6 @@ const LinkRender = ({
   pathName: string;
   closeTab: () => void;
 }) => {
-  const linkAppearAnimationVariant = {
-    initial: {
-      y: "30vh",
-      transition: {
-        duration: 0.5,
-        ease: [0.37, 0, 0.63, 1],
-      },
-    },
-    animate: {
-      y: 0,
-      transition: {
-        duration: 0.7,
-        ease: [0, 0.55, 0.45, 1],
-      },
-    },
-    exit: {},
-  };
   return (
     <motion.div variants={linkAppearAnimationVariant} onClick={closeTab}>
       <Link href={eachRoute.address}>
